Добавляет новый модуль и переносит в него соответствующие функции
