+++ source/js/main.js
@@ -1,7 +1,8 @@
 /* global _:readonly */
-import {setUserFormSubmit, onMapFiltersChange} from './form.js';
+import {setUserFormSubmit} from './form.js';
 import {fetchSimilarAds} from './api.js';
 import {renderSimilarPoints} from './map.js';
+import {onMapFiltersChange} from './similar-ads.js';
 import './photo.js';
 
 const RERENDER_DELAY = 500;
