+++ source/js/map.js
@@ -1,10 +1,10 @@
 /* global L:readonly */
 import {getCardElement} from './card.js';
+import {filterSimilarAds} from './similar-ads.js';
 
 const BASED_LOCATION_X = 35.6895000;
 const BASED_LOCATION_Y = 139.6917100;
 const MAP_ZOOM = 12;
-const COUNT_OF_SIMILAR_ADS = 10;
 
 const mainFormElement = document.querySelector('.ad-form');
 const mainFormFieldsetElements = mainFormElement.querySelectorAll('fieldset');
@@ -78,43 +78,6 @@ mainPinMarker.on('moveend', (evt) => {
   const {lat, lng} = evt.target.getLatLng();
   document.querySelector('#address').value = `${lat.toFixed(5)}, ${lng.toFixed(5)}`;
 });
-
-const filterSimilarAds = (list) => {
-  return list.filter(ad => {
-    const result = [];
-    const type = mapFiltersFormElement.querySelector('#housing-type').value;
-    if (type !== 'any') {
-      result.push(ad.offer.type === type)
-    }
-    const price = mapFiltersFormElement.querySelector('#housing-price').value;
-    if (price !== 'any') {
-      if (price === 'low') {
-        result.push(ad.offer.price < 10000)
-      }
-      if (price === 'middle') {
-        result.push( ad.offer.price > 10000 && ad.offer.price < 50000)
-      }
-      if (price === 'high') {
-        result.push(ad.offer.price > 50000)
-      }
-    }
-    const rooms = mapFiltersFormElement.querySelector('#housing-rooms').value;
-    if (rooms !== 'any') {
-      result.push(ad.offer.rooms === Number(rooms))
-    }
-    const guests = mapFiltersFormElement.querySelector('#housing-guests').value;
-    if (guests !== 'any') {
-      result.push(ad.offer.guests === Number(guests))
-    }
-    const features = mapFiltersFormElement.querySelectorAll('input[name="features"]');
-    for (let i = 0; i < features.length; i++) {
-      if (features[i].checked) {
-        result.push(ad.offer.features.indexOf(features[i].value) !== -1)
-      }
-    }
-    return result.every(el => el === true);
-  }).slice(0, COUNT_OF_SIMILAR_ADS);
-}
 
 // Выводит на карту метки похожих объявлений
 export const renderSimilarPoints = (ads) => {
